new prisma models Author & User
