+++ prisma/schema.prisma
@@ -11,19 +11,30 @@ datasource db {
 }
 
 model Post {
-  id        String   @id @default(auto()) @map("_id") @db.ObjectId
-  createdAt DateTime @default(now())
-  updatedAt DateTime @updatedAt
-  title     String
-  category  String
-  content   String
-  author    String
-  image     String
-  snippet   String
+  id         String   @id @default(auto()) @map("_id") @db.ObjectId
+  createdAt  DateTime @default(now())
+  updatedAt  DateTime @updatedAt
+  title      String
+  category   String
+  content    String
+  image      String
+  snippet    String
+  authorName String
+  author     Author?  @relation(fields: [authorId], references: [id])
+  authorId   String?  @db.ObjectId
 }
 
-model User {
-  id       String @id @default(auto()) @map("_id") @db.ObjectId
-  user     String
-  password String
+model Subscriber {
+  id            String   @id @default(auto()) @map("_id") @db.ObjectId
+  email         String   @unique
+  createdAt     DateTime @default(now())
+  verifiedEmail Boolean
+}
+
+model Author {
+  id    String @id @default(auto()) @map("_id") @db.ObjectId
+  name  String
+  bio   String
+  image String
+  posts Post[]
 }
